[fix] #70 MissionForChallengeVo attendanceCount status, result 조건 추가

+++ src/main/java/org/letscareer/letscareer/domain/mission/dto/response/MissionAdminResponseDto.java
@@ -16,7 +16,7 @@ public record MissionAdminResponseDto(
         Integer th,
         String missionType,
         MissionStatusType missionStatusType,
-        Integer attendanceCount,
+        Long attendanceCount,
         Long lateAttendanceCount,
         Long applicationCount,
         Integer score,
